Refactor denoising stages into dedicated modules
