+++ fastvideo/pipelines/stages/__init__.py
@@ -7,11 +7,11 @@ complete diffusion pipelines.
 """
 
 from fastvideo.pipelines.stages.base import PipelineStage
-from fastvideo.pipelines.stages.causal_denoising import CausalDMDDenosingStage
 from fastvideo.pipelines.stages.conditioning import ConditioningStage
 from fastvideo.pipelines.stages.decoding import DecodingStage
-from fastvideo.pipelines.stages.denoising import (DenoisingStage,
-                                                  DmdDenoisingStage)
+from fastvideo.pipelines.stages.denoising import (
+    CausalDMDDenosingStage, DenoisingStage, DmdDenoisingStage
+)
 from fastvideo.pipelines.stages.encoding import EncodingStage
 from fastvideo.pipelines.stages.image_encoding import (ImageEncodingStage,
                                                        ImageVAEEncodingStage)
